Finalizing ui, adding ability to hover over calendar events box listing in order to view full details on selected item. Added a tracker for hours spent in class and wages made, both for the week of the selected date.
Further, todo items of the same name and calendar events beginning at the same time are now disallowed from being created, with an error message prompting the user to change this before creating or editing listed items. Events ending before they begin should now be disallowed too.

Next Step: save user data

+++ TDEWork.cs
@@ -48,9 +48,9 @@
         {
             return this.StartTime.ToString("h:mm tt") + " - " + this.Name + " @ " + this.Company;
         }
-        public decimal GetPay()
+        public double GetPay()
         {
-            return _wage * Math.Round((decimal)(EndTime - StartTime).TotalHours, 2);
+            return (double)_wage * Math.Round((EndTime - StartTime).TotalHours, 2);
         }
         public double ShiftLength()
         {
